+++ src/units/heating-value.ts
@@ -1,6 +1,7 @@
 import * as UnitDivide from "../unit-divide";
+import { Kilojoule, Btu } from "./energy";
 import { KiloWattHour } from "./energy2";
-import { CubicMeter } from "./volume";
+import { CubicMeter, CubicFeet } from "./volume";
 
 export type HeatingValue = "HeatingValue";
 
@@ -11,4 +12,16 @@ export const KiloWattHourPerCubicMeter = UnitDivide.energyByVolume(
   "KiloWattHourPerCubicMeter",
   KiloWattHour,
   CubicMeter
+);
+
+export const KiloJoulePerCubicMeter = UnitDivide.energyByVolume(
+  "KiloJoulePerCubicMeter",
+  Kilojoule,
+  CubicMeter
+);
+
+export const BtuPerCubicFeet = UnitDivide.energyByVolume(
+  "BtuPerCubicFeet",
+  Btu,
+  CubicFeet
 );